chore: move actions service to new folder structure (#943)
* feat: move actions service

* fix: rename folder to be singular not plural

---------
