+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/actions/actions.ts
@@ -11,7 +11,7 @@ import {
   getActionCountInLast24Hours,
   getActionCountInLast7Days,
   getActionCountInLastHour,
-} from "@formbricks/lib/services/actions";
+} from "@formbricks/lib/action/service";
 import { getSurveysByActionClassId } from "@formbricks/lib/survey/service";
 import { AuthorizationError } from "@formbricks/types/v1/errors";
